+++ packages/extension/src/shared/multicall/index.ts
@@ -1,2 +1,1 @@
 export * from "./getMulticall"
-export * from "./getTokenBalance"
