+++ packages/extension/src/ui/features/accountTokens/TokenListItemContainer.tsx
@@ -1,17 +1,11 @@
-import { BigNumber } from "ethers"
-import { FC, useMemo } from "react"
+import { FC } from "react"
 
 import { Token } from "../../../shared/token/type"
-import { isNumeric } from "../../../shared/utils/number"
 import { withPolling } from "../../services/swr"
 import { Account } from "../accounts/Account"
 import { TokenListItem, TokenListItemProps } from "./TokenListItem"
 import { useTokenBalanceToCurrencyValue } from "./tokenPriceHooks"
-import { TokenDetailsWithBalance } from "./tokens.state"
-import {
-  TokenBalanceErrorMessage,
-  useTokenBalanceForAccount,
-} from "./useTokenBalanceForAccount"
+import { useTokenBalanceForAccount } from "./useTokenBalanceForAccount"
 
 export interface TokenListItemContainerProps
   extends Omit<TokenListItemProps, "currencyValue"> {
@@ -28,34 +22,20 @@ export const TokenListItemContainer: FC<TokenListItemContainerProps> = ({
   account,
   ...rest
 }) => {
-  const { data, isValidating } = useTokenBalanceForAccount(
-    {
-      token,
-      account,
-      shouldReturnError:
-        true /** using Suspense, causes error to be returned as `balance` instead of throwing */,
-    },
-    {
-      suspense:
-        true /** Suspense allows us to show an initial loader for all tokens */,
-      ...withPolling(30 * 1000) /** 30 seconds */,
-    },
-  )
-  const { tokenWithBalance, errorMessage } = useMemo(() => {
-    const tokenWithBalance: TokenDetailsWithBalance = {
-      ...token,
-    }
-    let errorMessage: TokenBalanceErrorMessage | undefined
-    if (isNumeric(data)) {
-      tokenWithBalance.balance = BigNumber.from(data)
-    } else {
-      errorMessage = data as TokenBalanceErrorMessage
-    }
-    return {
-      tokenWithBalance,
-      errorMessage,
-    }
-  }, [data, token])
+  const { tokenWithBalance, errorMessage, isValidating } =
+    useTokenBalanceForAccount(
+      {
+        token,
+        account,
+        shouldReturnError:
+          true /** using Suspense, causes error to be returned as `balance` instead of throwing */,
+      },
+      {
+        suspense:
+          true /** Suspense allows us to show an initial loader for all tokens */,
+        ...withPolling(60 * 1000) /** 30 seconds */,
+      },
+    )
 
   const currencyValue = useTokenBalanceToCurrencyValue(tokenWithBalance)
   const shouldShow =
